Rename FilterTagsRender to FilterTagsGroup
