+++ src/components/ViewPage.js
@@ -3,7 +3,7 @@ import { ReactComponent as RightArrow } from "../icons/arrow-right-thin.svg";
 import { ReactComponent as LeftArrow } from "../icons/arrow-left-thin.svg";
 import { ReactComponent as Progress } from "../icons/view-progress.svg";
 import { Link } from "react-router-dom";
-import FilterTagsRender from "./FilterTagsRender";
+import FilterTagsRender from "./FilterTagsGroup";
 
 export default function ViewPage({ restaurantData }) {
   const tags = ["habor view", "park view", "canal view", "alster lake view"];
